pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telemetryflow-sdk"
version = "1.1.2"
description = "TelemetryFlow Python SDK - OpenTelemetry-based observability SDK for metrics, logs, and traces"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
authors = [
    {name = "TelemetryFlow", email = "support@telemetryflow.id"}
]
maintainers = [
    {name = "TelemetryFlow", email = "support@telemetryflow.id"}
]
keywords = [
    "telemetry",
    "observability",
    "opentelemetry",
    "otel",
    "metrics",
    "logs",
    "traces",
    "monitoring",
    "apm",
    "telemetryflow"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]

dependencies = [
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "opentelemetry-semantic-conventions>=0.49b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "pre-commit>=3.8.0",
]
http = [
    "flask>=3.0.0",
    "fastapi>=0.115.0",
    "starlette>=0.41.0",
]
grpc = [
    "grpcio>=1.68.0",
]
restapi = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-limiter>=3.5.0",
    "flask-jwt-extended>=4.6.0",
    "flask-sqlalchemy>=3.1.0",
    "flasgger>=0.9.7",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://telemetryflow.id"
Documentation = "https://docs.telemetryflow.id"
Repository = "https://github.com/telemetryflow/telemetryflow-python-sdk"
Issues = "https://github.com/telemetryflow/telemetryflow-python-sdk/issues"
Changelog = "https://github.com/telemetryflow/telemetryflow-python-sdk/blob/main/CHANGELOG.md"

[project.scripts]
telemetryflow-gen = "telemetryflow.cli.generator:main"
telemetryflow-restapi = "telemetryflow.cli.generator_restapi:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["telemetryflow*"]

[tool.setuptools.package-data]
telemetryflow = ["py.typed", "cli/templates/**/*.tpl", "cli/templates/**/*.md.tpl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src/telemetryflow"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/cli/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
fail_under = 80

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true
exclude = [
    "tests/",
    "examples/",
]

[[tool.mypy.overrides]]
module = [
    "opentelemetry.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "ARG",   # flake8-unused-arguments
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["telemetryflow"]

[tool.black]
target-version = ["py312", "py313"]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["telemetryflow"]
skip_gitignore = true
